+++ R/as_tibble.R
@@ -1,12 +1,19 @@
+#' DEPRECATED: Coerce to tibble. Enable preserving row names when coercing matrix
+#' and time-series-like objects with row names.
+#'
+#' @description
+#'
 #' Coerce to tibble. Enable preserving row names when coercing matrix
 #' and time-series-like objects with row names.
+#'
+#' DEPRECATED: Use [timekit::tk_tbl()] instead.
 #'
 #' @param x A list, matrix, xts, zoo, timeSeries, etc object.
 #' @param preserve_row_names Used during coercion from matrix, xts, zoo,
 #' timeSeries, etc objects that have row names. When `TRUE`, creates
 #' a `row.names` column with names of rows as character class.
 #' @param ... Additional parameters passed to the appropriate
-#' [tibble::as_tibble()] function.
+#' [timekit::tk_tbl()] function.
 #'
 #' @return Returns a `tibble` object.
 #'
@@ -17,13 +24,15 @@
 #' When `preserve_row_names = TRUE` is specified, a new column,
 #' `row.names`, is created during object coercion as a character class.
 #'
-#' @seealso [as_xts()]
+#' @seealso
+#' * [tk_xts()] - Coercion to xts, replaces [tidyquant::as_xts()]
+#' * [tk_tbl()] - Coercion to tbl, replaces [tidyquant::as_tibble()]
+#' * [as_xts()] - Deprecated, use [tk_xts()] instead.
 #'
 #' @export
 #'
 #' @examples
 #' # Load libraries
-#' library(tidyverse)
 #' library(tidyquant)
 #'
 #' # Matrix coercion to tibble
@@ -38,85 +47,12 @@
 #'
 as_tibble <- function(x, preserve_row_names = FALSE, ...) {
 
-    warn <- FALSE
+    warning("The `tidyquant::as_tibble()` function is deprecated. Please use `timekit::tk_tbl()` instead.")
 
-    # Handle various time series objects
-    ret <- tryCatch({
-
-        if (xts::is.xts(x) ||
-            zoo::is.zoo(x) ||
-            timeSeries::is.timeSeries(x) ||
-            stats::is.ts(x)) {
-
-            matrix_to_tibble(as.matrix(x), preserve_row_names, ...)
-
-        } else if (tseries::is.irts(x)) {
-
-            matrix_to_tibble(x %>% xts::as.xts() %>% as.matrix(),
-                                    preserve_row_names, ...)
-
-        } else if (is.matrix(x)) {
-
-            matrix_to_tibble(x, preserve_row_names, ...)
-
-        } else {
-
-            if (preserve_row_names) warn <- TRUE
-            tibble::as_tibble(x, ...)
-
-        }
-
-    }, error = function(e) {
-
-        warning(paste0("Error at ", x,
-                       ": Could not convert to tibble. Check input."))
-        NA
-
-    })
-
-    if (warn) {
-        warning(paste0("\nWarning at input ", x, ":",
-                       "\nPreserving row names not used for this object class.",
-                       " Object was otherwise converted to tibble successfully."))
-    }
-
-    ret
+    timekit::tk_tbl(data = x, preserve_index = preserve_row_names, silent = TRUE, ...)
 
 }
 
-# UTILITY FUNCTIONS ----
 
-# matrix to tibble conversion:
-#     allow preserving rownames when converting to tibble
-matrix_to_tibble <- function(x, preserve_row_names, ...) {
-
-    if (!is.matrix(x)) stop("Error: `x` is not a matrix object.")
-
-    if (preserve_row_names == TRUE) {
-
-        row.names <- rownames(x)
-
-        # Detect if row.names exist beyond sequential 1:nrow(x) or null value
-        if (!identical(row.names, 1:nrow(x) %>% as.character()) &&
-            !is.null(row.names)) {
-
-            dplyr::bind_cols(
-                tibble::tibble(row.names),
-                tibble::as_tibble(x, ...)
-            )
-
-        } else {
-
-            warning(paste0("Warning: No row names to preserve. ",
-                           "Object otherwise converted to tibble successfully."))
-            tibble::as_tibble(x, ...)
-        }
-
-    } else {
-
-        tibble::as_tibble(x, ...)
-
-    }
-}